Split into component files Started on task-bar Changed el/this.el to win/this.window
