+++ app/index.html
@@ -9,7 +9,7 @@
   <script defer src="https://use.fontawesome.com/releases/v5.0.6/js/all.js"></script>
 </head>
 <body>
-  <div id="window-template">
+  <div id="window-template" class="template">
       <div class="window">
         <div class="window__button close"><i class="fas fa-times"></i></div>
         <div class="window__title"></div>
@@ -31,6 +31,8 @@
         <div class="window__button size">&#x25E2;</div>
       </div>
   </div> 
+
+  <div id="task-template" class="template"><div class="task">window</div></div>
 
   <div class="wrapper">
       <div id="menu-bar"></div>
