+++ app/index.scss
@@ -5,7 +5,7 @@ html {
   box-sizing: inherit;
 }
 
-#window-template {
+.template {
   display: none;
 }
 
@@ -26,7 +26,7 @@ body {
   grid-template-rows: 1.5rem 1fr 2rem;
 }
 
-#menu-bar, #task-bar {
+#menu-bar {
   z-index: 1000;
 }
 
@@ -35,84 +35,9 @@ body {
   border-bottom: black solid 1px;
 }
 
-#task-bar {
-  background: silver;
-  border-top: black solid 1px;
-}
-
 #desktop {
   background: hsl(120, 20%, 60%);
 }
 
-.window {
-  box-shadow: 5px 10px 10px rgba(0,0,0,0.3), -2px -2px 20px rgba(0,0,0,0.3);
-  position: absolute;
-  display: grid;
-  border: 1px solid black;
-  grid-template-columns: 1.5rem 1fr 1.5rem 1.5rem;
-  grid-template-rows: 1.5rem 1.5rem 1fr 1.5rem 1.5rem; 
-  font-size: 12px;
-  line-height: 10px;
-  font-family: monospace;
-  
-  > * {
-    display: grid;
-    justify-content: center;
-    align-content: center;
-    border: 1px solid black;
-    cursor: default;
-  }
-  
-  &__content {
-    grid-row: span 3;
-    grid-column: span 3;
-    background: white;
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-template-rows: 1fr;
-    overflow: hidden;
-    &__iframe, &__cover, &__loading {
-      grid-column: 1 / span 1;
-      grid-row: 1 / span 1;
-    }
-    &__iframe {
-      min-width: 0;
-      min-height: 0;
-      display: none;
-      border: none;
-    }
-    &__cover {
-      display: none;
-    }
-    &__loading {
-      display: grid;
-      justify-content: center;
-      align-content: center;
-    }
-  }
-  
-  &__title {
-    background: lightsteelblue;
-    overflow: hidden;
-    white-space: nowrap;
-  }
-  
-  &__scroll-bar {
-    background: lightgrey;
-  }
-  
-  &__button {
-    background: silver;
-    &.size.disabled {
-      color: slategray;
-    }
-  }
-  
-  &.changing {
-    border: 1px solid white;
-  }
-
-  &.maximising {
-    transition: 0.1s all; 
-  }
-}
+@import './window.scss';
+@import './task-bar.scss';
