+++ app/task-bar.scss
@@ -1,0 +1,26 @@
+#task-bar {
+  z-index: 1000;
+  background: lightgray;
+  border-top: black solid 1px;
+  display: grid;
+  grid-template-columns: repeat(auto-fit,minmax(1rem, 1fr));
+  grid-column-gap: 3px;
+  padding: 3px;
+  font-size: 0.75rem;
+}
+
+.task {
+  min-width: 0;
+  display:grid;
+  align-content: center;
+  justify-content: center;
+  box-shadow: inset -1px -1px 1px #000, inset 1px 1px 1px #fff;
+  &.active {
+    background: silver;
+    box-shadow: inset 1px 1px 1px #000, inset -1px -1px 1px #fff;
+  }
+  &.minimised {
+    color: gray;
+    font-style: italic;
+  }
+}
