+++ app/window.scss
@@ -1,0 +1,76 @@
+.window {
+  box-shadow: 5px 10px 10px rgba(0,0,0,0.3), -2px -2px 20px rgba(0,0,0,0.3);
+  position: absolute;
+  display: grid;
+  border: 1px solid black;
+  grid-template-columns: 1.5rem 1fr 1.5rem 1.5rem;
+  grid-template-rows: 1.5rem 1.5rem 1fr 1.5rem 1.5rem; 
+  font-size: 12px;
+  line-height: 10px;
+  font-family: monospace;
+  
+  > * {
+    display: grid;
+    justify-content: center;
+    align-content: center;
+    border: 1px solid black;
+    cursor: default;
+  }
+  
+  &__content {
+    grid-row: span 3;
+    grid-column: span 3;
+    background: white;
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr;
+    overflow: hidden;
+    &__iframe, &__cover, &__loading {
+      grid-column: 1 / span 1;
+      grid-row: 1 / span 1;
+    }
+    &__iframe {
+      min-width: 0;
+      min-height: 0;
+      display: none;
+      border: none;
+    }
+    &__cover {
+      display: none;
+    }
+    &__loading {
+      display: grid;
+      justify-content: center;
+      align-content: center;
+    }
+  }
+  
+  &__title {
+    background: lightsteelblue;
+    overflow: hidden;
+    white-space: nowrap;
+  }
+  
+  &__scroll-bar {
+    background: lightgrey;
+  }
+  
+  &__button {
+    background: silver;
+    &.size.disabled {
+      color: slategray;
+    }
+  }
+  
+  &.changing {
+    border: 1px solid white;
+  }
+
+  &.maximising {
+    transition: 0.1s all; 
+  }
+
+  &.minimised {
+    display: none;
+  }
+}
