+++ package.json
@@ -13,7 +13,8 @@
     "test": "echo \"No tests. Sorry\""
   },
   "dependencies": {
+    "node-sass": "^4.7.2",
     "parcel-bundler": "^1.5.1",
-    "node-sass": "^4.7.2"
+    "uuid": "^3.2.1"
   }
 }
